requirements.txt
requests
beautifulsoup4
pandas
googlesearch-python
pytest